Add RetrieveSpecificAnalysis command
Given a 3 number concatenated id, this command will return the analysis
for that analysis. It has been tested and seems to work well. It is
still using fake data, but once again thats only because the other parts
of the project that generate real data have not been integrated yet.