Make retrieveAllSummaries return the results and main print the results
Before this commit, the Query() object printed the results to the interface.
However, this is non modularized because main is supposed to interact with the
user. Also the results of the retrieveAllSummaries is needed in other methods.
So moving the printing results to the main interface is a better design choice.