Add basic README.md file
README will be updated later. This will be the initial readme
when pushing it to GitHub for safery purposes. Also note that because
the SQL isnt hosted that the app currently wont work from a fork or
similar command. It is necessary to fix that within the next few
commits